Rename the Database namespace to Outbo
