+++ Libraries/CO.CDP.MQ/Outbox/IOutboxMessageDbContext.cs
@@ -1,6 +1,6 @@
 using Microsoft.EntityFrameworkCore;
 
-namespace CO.CDP.MQ.Database;
+namespace CO.CDP.MQ.Outbox;
 
 public interface IOutboxMessageDbContext
 {
